refactor: consolidate ticket functionality into main components with toggle
